requirements.txt
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.3.2
ecdsa==0.14.1
fake-useragent==0.1.11
fastapi==0.63.0
h11==0.12.0
idna==2.10
lxml==4.6.2
mysqlclient==2.0.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
python-dotenv==0.15.0
python-jose==3.2.0
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
wincertstore==0.2
